refactor: replace email preference API route with Astro actions
Replace manual fetch calls to /api/notifications/email-preference with
type-safe Astro actions (getEmailPreference, setEmailPreference).
Removes the API route in favor of src/actions/index.ts.
